requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
langchain>=0.0.350
langchain-openai>=0.0.2
openai>=1.3.0
faiss-cpu>=1.7.4
pypdf>=3.17.0
requests>=2.31.0
pydantic>=2.5.0
docx2txt>=0.8
unstructured>=0.11.0
python-dotenv>=1.0.0